Fetch gerrit patches in ABE's lib/checkout.sh rather than in jenkins.sh.
The gerrit patches are in refs/changes/*, which are not fetched
by default by git clone.  Since we are putting more and more stuff
to gerrit, it makes sense for us to fetch these commits in ABE.

The extra refs are going to be present in the reference git repos,
so overhead will be negligible.

Change-Id: I3043decb858231a3fcf2bf8a40ed12f4550e8739