Fix git pull failure when building from tag.
If a component is configured to build a tag rather than a branch, then
checkout() causes this error:
    You are not currently on a branch. Please specify which
    branch you want to merge with. See git-pull(1) for details.

Since commit 04adeb37 (Report unhandled errors), this causes the build
to fail, as it is treated as an unhandled error. The fix is to only
try git pull when an upstream branch is configured.

Change-Id: Id52314ba4c6f55705040d6eb08422ba0d8f38cd5