When updating a branch, make sure the branch is checked out.
When doing --build all, checkout is called for both gcc stage 1
and stage 2. Since the checkout of stage 1 sets
gcc_component_revision, the checkout of stage 2 is performed as
though the user requested a checkout of a specified revision. This
means that the srcdir is on the local_${revision} branch, not the
user's requested branch. This causes a git pull in any subsequent
build to fail.

This fixes:
 ERROR (#321): checkout (Error during update_checkout_branch.)
when doing a second build of the same branch.

This wasn't caught by testing because "--checkout all" only checks
out gcc once, so the abe-tests-checkout does not trigger this bug.

Change-Id: Ie32d252112488de47390985d683b68d3254e7895