Revert gdbserver commit fd3907fb6c8da336c1ff948bc7e1eccd741f4a24
I forgot to revert the GDB server commit with the last commit
and just reverted GDB proper.

Reverting the gdbserver update to 7.12-branch because there
are regressions in the GDB 7.12 release for big-endian targets.
This commit reverts only the GDB 7.11 -> GDB 7.12, and not the
addition of CXXFLAGS=--sysroot=${sysroots} for the GDB build.

    commit fd3907fb6c8da336c1ff948bc7e1eccd741f4a24
    Author: Christophe Lyon <christophe.lyon@linaro.org>
    Date:   Mon Oct 10 21:18:07 2016 +0200

        Upgrade gdb/gdbserver to 7.12-branch.

Change-Id: I654839b663067d42ed7a27e5230f482e5b999926