configure: Use the total number of CPUs.
Instead of just the online ones. This is because some boards
activate CPUs on-demand, so getconf _NPROCESSORS_ONLN could
return '1' although other CPUs are offline.

This will speed up builds in such cases, because the
build started by abe will increase the load on the
machine, which in turn will make the system activate
more CPUs.

Change-Id: Icf9a7bbe304af8a37dfc79f7fa7f7119d2aa4883