Add CPPFLAGS to gdbserver configure line.
Since gdbserver is built before the INSTALL_SYSROOT step, it
is necessary to use --sysroot when invoking the compiler. We
already set CFLAGS and CXXFLAGS, but also need to set CPPFLAGS
so that the configure script can correctly find header files for
testing platform features.

This fixes the behaviour of the configure script. On Arch Linux,
the gdbserver configure script aborts with
  configure: error: C preprocessor "/lib/cpp" fails sanity check
On other distros which supply a link to the host system's C
preprocessor in /lib/cpp, this patch ensures that the cross
toolchain's CPP is used instead.

Change-Id: I0fddc6ea8be0ac553a84c16470aca6b66f955c50