Unify mingw and linux manifests.
In order to unify the manifests generated for Windows and Linux
hosts, add mingw_only flag to all components, and an optional
mingw_extraconf flag.

Expat and Python are added to the list of components in Linux
builds but are skipped during building on Linux hosts.

The configuration for gdb is updated to use mingw_extraconf.

Bump manifest version to 1.4, and fix up manifest parsing
accordingly; support manifest versions <= 1.3 by:
. automatically setting mingw_only flag for all components
. setting mingw_extraconf to "" for all components
. only building components which are described in the manifest
  (so that expat/python are not built or checked out on Linux
  hosts)

This means that manifest versions 1.1, 1.2 and 1.3 are still
tied to the host that they were created for, but new manifests
will be identical for Linux and Windows hosts.

Change-Id: I56085ed24a11e20ede0be6b96d5ec9125d7e7e45