abe.sh: Accept --set buildconfig=XXX without bootstrap if no build requested.
gcc's --with-build-config=XXX option is only supported for bootstrap builds.
However, if we don't request a build, there is no need to exit early if
--set buildconfig=XXX was used without --enable bootstrap.

This is needed when Abe is called from jenkins.sh, which does a first
pass with --checkout all (and without --build). This first pass
includes modifier flags such as --set buildconfig=XXX, so we want
it to complete without complaining.

Change-Id: If5a2dac44e0114495e14c6a0e061ad08b32563c2