Fix newlib builds from manifests.
Since commit 72071d0268b1742b36fa7767d407b2b9a96c048b (Unify mingw
and linux manifests), Abe fails to build newlib when it is the
selected C library in a manifest.

This happens because newlib doesn't match the "virtual" package
name 'libc' in the list of built components. This was missed by
the gerrit testing because it builds only linux targets and "glibc"
matches the grep for "libc".

The fix is to rewrite "libc" as ${clibrary} before trying to match
package names.

Change-Id: I2d165ea5a89e6b00766cb6e78d6a0a4b13310c49