Record location of .sum files as artifacts.
In order to eliminate the use of find in jenkins-scripts and the
Jenkins job yaml files, add the location of .sum files to the
artifacts list.

Unfortunately, we still have to do a 'find' here for now, as the
projects themselves do not place the sum files in predictable
locations. This still has the benefit that the find is only executed
in build dirs where the sum files are expected to be found.

Change-Id: Id876be46e3ffdc5bda97c32718c0f5c71536a1ba